将id类型改为Long型

+++ app/src/main/java/com/xiaolian/amigo/data/network/model/dto/request/BindResidenceReq.java
@@ -9,6 +9,6 @@ import lombok.Data;
  */
 @Data
 public class BindResidenceReq {
-    private Integer id;
-    private Integer residenceId;
+    private Long id;
+    private Long residenceId;
 }
