+++ app/src/main/java/com/xiaolian/amigo/data/network/model/dto/request/PersonalUpdateReqDTO.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 public class PersonalUpdateReqDTO {
 
-    private Integer residneceId;
+    private Long residneceId;
     private String nickName;
     private String pictureUrl;
-    private Integer schoolId;
+    private Long schoolId;
     private Integer sex;
 }
